update for jobs and clean up
